add authentication. closes #4

+++ internal/list/item.go
@@ -5,11 +5,11 @@ import (
 )
 
 type Item struct {
-	ID          uint          `json:"id" gorm:"primary_key"`
+	ID          uint64        `json:"id" gorm:"primary_key"`
 	Description string        `json:"description" gorm:"NOT NULL"`
 	Duration    time.Duration `json:"duration,omitempty"`
 	IsDone      bool          `json:"is_done" gorm:"type:boolean"`
-	Weight      uint          `json:"weight"`
+	Weight      uint64        `json:"weight"`
 
 	CreatedAt time.Time  `json:"-"`
 	UpdatedAt time.Time  `json:"-"`
