+++ internal/list/repository.go
@@ -2,8 +2,8 @@ package list
 
 type Repository interface {
 	Insert(item Item) error
-	FindById(id uint) (Item, error)
+	FindById(id uint64) (Item, error)
 	Update(item Item) error
-	Delete(id uint) error
+	Delete(id uint64) error
 	FindByIsDone(isDone bool) ([]Item, error)
 }
